[refactor] RemoteNickNameSetUpDataSourceImpl 수정 (#131)
