+++ android/app/src/main/java/com/teamtripdraw/android/TripDrawApplication.kt
@@ -18,14 +18,12 @@ class TripDrawApplication : Application() {
     private fun initContainer() {
         localPreferenceContainer = LocalPreferenceContainer(applicationContext)
         localDataSourceContainer = LocalDataSourceContainer(localPreferenceContainer)
-        retrofitContainer = RetrofitContainer(localDataSourceContainer.localUserIdentifyInfoDataSource)
+        retrofitContainer =
+            RetrofitContainer(localDataSourceContainer.localUserIdentifyInfoDataSource)
         serviceContainer = ServiceContainer(retrofitContainer)
-        remoteDataSourceContainer = RemoteDataSourceContainer(serviceContainer)
-        repositoryContainer = RepositoryContainer(
-            localDataSourceContainer,
-            remoteDataSourceContainer,
-            retrofitContainer
-        )
+        remoteDataSourceContainer = RemoteDataSourceContainer(serviceContainer, retrofitContainer)
+        repositoryContainer =
+            RepositoryContainer(localDataSourceContainer, remoteDataSourceContainer)
     }
 
     companion object DependencyContainer {
