+++ android/app/src/main/java/com/teamtripdraw/android/data/repository/NicknameSetupRepositoryImpl.kt
@@ -2,56 +2,19 @@ package com.teamtripdraw.android.data.repository
 
 import com.teamtripdraw.android.data.dataSource.nicknameSetup.NicknameSetupDataSource
 import com.teamtripdraw.android.data.dataSource.userIdentifyInfo.UserIdentifyInfoDataSource
-import com.teamtripdraw.android.data.httpClient.dto.failureResponse.NicknameSetupFailureResponse
-import com.teamtripdraw.android.domain.exception.DuplicateNickNameException
-import com.teamtripdraw.android.domain.exception.InvalidNickNameException
 import com.teamtripdraw.android.domain.repository.NicknameSetupRepository
-import com.teamtripdraw.android.support.framework.data.getParsedErrorBody
-import okhttp3.ResponseBody
-import retrofit2.Retrofit
 
 class NicknameSetupRepositoryImpl(
     private val localUserIdentifyInfoDataSource: UserIdentifyInfoDataSource.Local,
-    private val remoteNicknameSetupDataSource: NicknameSetupDataSource.Remote,
-    private val retrofit: Retrofit
+    private val remoteNicknameSetupDataSource: NicknameSetupDataSource.Remote
 ) :
     NicknameSetupRepository {
-    override suspend fun setNickname(nickname: String): Result<Long> {
-        return remoteNicknameSetupDataSource.setNickname(nickname)
-            .process(failureListener = this::setNickNameFailureListener) { body, headers ->
-                Result.success(body.memberId)
-            }
-    }
-
-    private fun setNickNameFailureListener(code: Int, errorBody: ResponseBody?): Result<Nothing> {
-        // todo error sealed class화 시키기
-        if (code == 409) {
-            val message =
-                retrofit.getParsedErrorBody<NicknameSetupFailureResponse>(errorBody)?.message
-            return Result.failure(
-                DuplicateNickNameException(
-                    message ?: DEFAULT_DUPLICATE_NICKNAME_EXCEPTION_MESSAGE
-                )
-            )
-        }
-        // 중복 닉네임을 제외한 오류들은 code400으로 분기되어있다.(#43 참고)
-        return Result.failure(
-            InvalidNickNameException(
-                DEFAULT_INVALID_NICKNAME_EXCEPTION_MESSAGE
-            )
-        )
-    }
+    override suspend fun setNickname(nickname: String): Result<Long> =
+        remoteNicknameSetupDataSource.setNickname(nickname)
 
     override suspend fun getNickname(nicknameId: Long): Result<String> =
         // todo #107 기준 로그인이 구현 안 되어 있는 문제로 임시로 닉네임을 통한 인증 상태 추후 로직 변경 필수(이슈 참고)
-        remoteNicknameSetupDataSource.getNickname(nicknameId).process { body, headers ->
-            localUserIdentifyInfoDataSource.setIdentifyInfo(body.nickname)
-            Result.success(body.nickname)
+        remoteNicknameSetupDataSource.getNickname(nicknameId).onSuccess { nickname ->
+            localUserIdentifyInfoDataSource.setIdentifyInfo(nickname)
         }
-
-    companion object {
-        private const val DEFAULT_DUPLICATE_NICKNAME_EXCEPTION_MESSAGE = "중복된 닉네임 입니다."
-        private const val DEFAULT_INVALID_NICKNAME_EXCEPTION_MESSAGE =
-            "공백 문자 혹은 10자를 초과한 부적절한 닉네임 입니다."
-    }
 }
